+++ crates/blockifier/src/execution.rs
@@ -1,4 +1,3 @@
-pub mod bouncer;
 pub mod call_info;
 pub mod common_hints;
 pub mod contract_address;
