+++ crates/blockifier/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod abi;
-pub mod block;
 pub mod blockifier;
 pub mod context;
 pub mod execution;
